+++ act/nari/ext/act/parser/__init__.py
@@ -1,11 +1,8 @@
 """Just a bunch of helper methods to spit out events from the ACT log"""
-from datetime import datetime
-from hashlib import md5, sha256
-from enum import IntEnum, Enum
+from enum import IntEnum
 
 from nari.types.event import Event
 from nari.types import Timestamp
-from nari.types.event.limitbreak import LimitBreak
 
 from nari.ext.act.parser.metadata import version_from_logline, config_from_logline
 from nari.ext.act.parser.zone import zonechange_from_logline
@@ -26,17 +23,7 @@ from nari.ext.act.parser.waymark import waymark_from_logline
 from nari.ext.act.parser.party import partylist_from_logline
 from nari.ext.act.parser.effectresult import effectresult_from_logline
 from nari.ext.act.parser.cast import startcast_from_logline, stopcast_from_logline
-from nari.ext.act.exceptions import InvalidActChecksumAlgorithm
 from nari.ext.act.types import ActIdMapping
-
-DEFAULT_DATE_FORMAT: str = '%Y-%m-%dT%H:%M:%S.%f%z'
-
-# pylint: disable=invalid-name
-class ActLogChecksumType(Enum):
-    """List of hashsum algorithms used by different ACT versions"""
-    MD5 = "md5"
-    SHA256 = "sha256"
-# pylint: enable=invalid-name
 
 
 # pylint: disable=invalid-name
@@ -89,29 +76,6 @@ class ActEventType(IntEnum):
         """Returns True if the name is in the enum"""
         return name in cls.__members__.keys() # pylint: disable=consider-iterating-dictionary
 # pylint: enable=invalid-name
-
-def date_from_act_timestamp(datestr: str) -> Timestamp:
-    """Parse timestamp from ACT log into a Timestamp
-    Look, this is dirty. This is wrong. Please someone find a better way to do this.
-    """
-    return int(datetime.strptime(f'{datestr[:26]}{datestr[-6:]}', DEFAULT_DATE_FORMAT).timestamp() * 1000)
-
-def validate_checksum(line: str, index: int, algo: ActLogChecksumType = ActLogChecksumType.SHA256) -> bool:
-    """Validates an ACT log line
-    Given some line 1|foo|bar|baz|a823425f532c540667195f641dd3649b, and an index of 1, then the md5sum of
-    1|foo|bar|baz|1 (where 1 is the index) should be a823425f532c540667195f641dd3649b (which is the checksum value)
-    """
-    parts = line.split('|')
-    check_hash = parts[-1]
-    to_hash = f'{"|".join(parts[:-1])}|{index}'.encode('utf-8')
-
-    match algo:
-        case ActLogChecksumType.MD5:
-            return md5(to_hash).hexdigest() == check_hash
-        case ActLogChecksumType.SHA256:
-            return sha256(to_hash).hexdigest()[:16] == check_hash
-        case _:
-            raise InvalidActChecksumAlgorithm(f'Unexpected checksum algorithm: {algo}. Expected one of MD5 and SHA256.')
 
 # pylint: disable=unused-argument
 def noop(timestamp: Timestamp, params: list[str]) -> Event:
